Pull up serial mock helper into a separate module.
