+++ tests/pygatt_tests.py
@@ -4,7 +4,6 @@ from binascii import unhexlify
 from mock import patch
 from nose.tools import eq_, assert_in
 import platform
-import Queue
 import unittest
 from struct import pack
 import threading
@@ -12,44 +11,7 @@ import time
 
 from pygatt.backends import BGAPIBackend
 
-
-class SerialMock(object):
-    """
-    Spoof a serial.Serial object.
-    """
-    def __init__(self, port, timeout):
-        self._isOpen = True
-        self._port = port
-        self._timeout = timeout
-        self._output_queue = Queue.Queue()
-        self._active_packet = None
-        self._expected_input_queue = Queue.Queue()
-
-    def open(self):
-        self._isOpen = True
-
-    def close(self):
-        self._isOpen = False
-
-    def write(self, input_data):
-        pass
-
-    def read(self):
-        if self._active_packet is None:
-            try:
-                self._active_packet = self._output_queue.get_nowait()
-            except Queue.Empty:
-                # When no bytes to read, serial.read() returns empty byte string
-                return b''
-        read_byte = self._active_packet[0]
-        if len(self._active_packet) == 1:  # we read the last byte
-            self._active_packet = None
-        else:
-            self._active_packet = self._active_packet[1:]
-        return read_byte
-
-    def stage_output(self, next_output):
-        self._output_queue.put(next_output)
+from .serial_mock import SerialMock
 
 
 class BGAPIBackendTests(unittest.TestCase):
